Add discoverable option to visibility
A file that is discoverable shows up in search results.  Its metadata
can be viewed by anyone, but the asset itself cannot be.

The thumbnail is not currently viewable in the search results.  That
will be addressed by the next commit.

We make the assumption that a discoverable asset should be readable by
any registered user; that may not always be a valid assumption, but
it’s fine for now.

This is almost certainly not the right way to implement this setting.
Discoverability and readability are really orthogonal concerns.
However, doing that will be a bigger deal involving front-end JS work,
because some combinations of discover and read permissions don’t make
sense.  This is a reasonable compromise for now.