Make date_created a required field
I needed to add the `required` attribute to the form helper as well as
adding to `required_fields` in `FileEditForm`.

The form helper is using the Rails `date_field` helper rather than the
SimpleForm version.  That means that the `aria-required` attribute is
not being set as it is for other form fields.  I tried to switch to the
SimpleForm variant, but it blew up because `date_created` is actually a
String, not a Date.  I gave up on this for now.