Validate aliases
We don’t want chains of aliases where A is an alias of B which is an
alias of C.  Rather, both A and B should be aliases of C.

We need to check for two things when assigning a canonical_venue to
another venue (or adding an alias to a venue):
1. The canonical venue being assigned can’t be an alias.
2. The venue that is becoming an alias can’t have aliases of its own.

In order to implement these checks, we need the reverse relationship so
a venue knows if it has aliases.  We use `dependent: nullify` so that
if a canonical venue is destroyed, its aliases become canonical venues.