Use our branch of hydra-head
We’ve submitted a PR for adding discover abilities to
hydra-access-controls (part of hydra-head).  I made a new branch,
`osf_integration`, off of v9.2.2 and then cherry-picked the PR commit
into that branch.

This commit removes the local versions of the PR code.

It also makes use of one of the methods added as part of the PR in
`solr_document.rb`.  This makes our implementation of `discover_groups`
more like Sufia’s implementation of `read_groups` and friends.