Allow batch edits to work with has_year_only checkbox
When doing batch edits, the model for the form is a `GenericFile` that
is not wrapped in a presenter as with normal edits and uploads.  As a
result, the `has_year_only` method from the presenter was not available
in that context, resulting in an error.

I moved the method over to the model and added a table-driven test.

I also realized that I could rename the method to `has_year_only?`
instead of `has_year_only` and use that new name in the view as well.
This addresses an earlier annoyance.

The batch edit view around these two fields and the checkbox is not
optimal and may need to be addressed.