Ensure production credits selectors display correctly
In the bulk upload form, the production credits selectors are initially
hidden.  Because of a [known bug in
chosen.js](https://github.com/harvesthq/chosen/issues/92), that causes
the Chosen widget to have a width of 0px.

The [Chosen
documentation](https://harvesthq.github.io/chosen/options.html)
indicates that you must specify a width explicitly in this case.

Based on the styling of the underlying select element, I chose to use
`width: “100%”`, and that seems to work quite well.