Add item detail divs after each row
The way this is done is to remove all of the extra layout elements
(`of-*` classes and detail divs) before performing layout, and then
adding them back as part of layout.  This is necessary because we
dynamically load more images and want to re-layout the grid when we do.

This approach has two implications:
* The more images we load, the slower this gets.  We could get smarter
by only laying out the last old row and all of the new cells, but
that’s harder and may not be necessary.

* Because we remove all of the detail divs, we will lose any expanded
div when we load more images.  If we’re not ok with this, we’ll have to
do something to remember what element was expanded and then restore it
after the re-layout.