Start handling clicks on cells
For now, just highlight the selected item with a red border.  Selecting
a cell deselects all other cells, as does hitting the Load More button.
 Clicking the selected cell deselects it.

To make this work, FlowGrid now gets initialized once and then we call
`relayout()` on it as necessary.  FlowGrid is also passed its root
element so it is scoped.  This makes it cleaner, and also possible to
have multiple independent FlowGrids on a page.